Feat/ldax control block (#38)
* rebased with dev


* ruff check


* commit for testing PR comments 1


* leonard call 1


* clang and casting done correctly


* removed test cases


* made changes leonard suggested 2


* removed one test case 1


* removed p_next_ldax since its not being used anywhere


---------

+++ cbxp/control_blocks/control_block.cpp
@@ -195,7 +195,7 @@ bool ControlBlock::compare(const nlohmann::json& json_value,
   uint64_t value_uint;
 
   if (json_value.is_number()) {
-    value_uint = json_value.get<int>();
+    value_uint = json_value.get<uint64_t>();
   } else {
     value_str       = json_value.get<std::string>();
     value_is_string = true;
